fix: add BulletinModel
